refactor(catchup): extract CatchUpContent for reuse on the home
